pyproject.toml
[project]
name = "sb-notify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.6.0",
    "django>=5.1.5",
    "pillow>=11.1.0",
    "psycopg[binary,pool]>=3.2.4",
    "django-unfold>=0.48.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.9.4",
]

[tool.ruff]
exclude = ["**/.venv/**", "**/migrations/**"]
lint.select = ["ALL"]
lint.ignore = ["D203", "D213", "ARG001", "S101", "RUF002", "RUF001", "D104", "RUF012",
            "ANN001", "ANN201", "N806", "D103", "D106", "D101", "D100"]
line-length = 79

[tool.ruff.lint.mccabe]
max-complexity = 3

[tool.black]
line-length = 79
